# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  chat_sdk.SendVerifyCodeReq:
    properties:
      identifier:
        description: 手机号或邮箱
        example: "13800138000"
        type: string
      purpose:
        description: register/forgot_password
        example: register
        type: string
    required:
    - identifier
    - purpose
    type: object
...
  chat_sdk.UpdateUserAvatarReq:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
    required:
    - avatar
    type: object
  chat_sdk.UpdateUserPasswordReq:
    properties:
      new_password:
        example: "123456"
        type: string
    required:
    - new_password
    type: object
...
  response.Response:
    properties:
      code:
        description: 业务状态码
        example: 0
        type: integer
      data:
        description: 响应数据
        type: object
      msg:
        description: 提示消息
        example: success
        type: string
    type: object
...
  service.ForgotPasswordReq:
    properties:
      code:
        type: string
      identifier:
        description: phone/email/username(这里允许 username，但更推荐 phone/email)
        type: string
      new_password:
        type: string
    type: object
...
  service.LoginReq:
    properties:
      account:
        description: username/phone/email
        type: string
      code:
        description: 验证码（可选：与 password 二选一）
        type: string
      password:
        description: plaintext（可选：与 code 二选一）
        type: string
    type: object
  service.LoginResp:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/service.UserDTO'
    type: object
...
  service.RegisterReq:
    properties:
      code:
        type: string
      email:
        description: phone/email 二选一
        type: string
      password:
        type: string
      phone:
        description: phone/email 二选一
        type: string
      username:
        type: string
    type: object
...
  service.SendCodeResult:
    properties:
      code:
        description: 是否返回由上层/调用方决定；这里总是返回，便于集成发送通道与测试
        type: string
      ttl_seconds:
        type: integer
    type: object
...
  service.UpdateUserReq:
    properties:
      birthday:
        type: string
      email:
        type: string
      gender:
        type: integer
      nickname:
        type: string
      phone:
        type: string
      signature:
        type: string
    type: object
...
  service.UserDTO:
    properties:
      avatar:
        type: string
      birthday:
        type: string
      created_at:
        type: string
      email:
        type: string
      gender:
        type: integer
      group_nickname:
        description: 我在该群里的昵称（群成员/会话列表可用）
        type: string
      id:
        type: integer
      last_active_at:
        type: string
      last_login_at:
        type: string
      nickname:
        type: string
      online_status:
        type: integer
      phone:
        type: string
      remark:
        description: 好友备注（仅在好友/私聊场景有意义）
        type: string
      room_account:
        description: 私聊房间对外号（与该好友的会话）
        type: string
      room_id:
        description: 私聊房间ID（与该好友的会话）
        type: integer
      signature:
        type: string
      uid:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:6789
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/cydxin/chat-sdk/issues
  description: |-
    即时通讯 SDK 的 RESTful API 文档，包含用户管理、消息、房间、好友等模块

    ## 业务状态码说明
    | Code | 说明 |
    |------|------|
    | 0 | 成功 |
    | 10001 | 参数错误 |
    | 10002 | 用户不存在 |
    | 10003 | 密码错误（登录失败） |
    | 10004 | Token 无效 |
    | 10005 | 权限不足 |
    | 99999 | 内部错误 |

    ## HTTP 状态码说明
    - **200**: 业务请求成功（根据 response.code 判断业务状态）
    - **401**: 认证失败（未登录/Token 无效/登录失败）
    - **403**: 权限不足
    - **500**: 服务器内部错误

    ## 响应格式
    所有接口统一返回格式：
    ```json
    {
    "code": 0,
    "msg": "success",
    "data": {}
    }
    ```
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/cydxin/chat-sdk
  title: Chat SDK API
  version: "1.0"
paths:
...
  /user/avatar:
    post:
      consumes:
      - application/json
      description: 更新当前用户头像 URL
      parameters:
      - description: 头像更新
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.UpdateUserAvatarReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的用户信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.UserDTO'
              type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新用户头像
      tags:
      - 用户
  /user/code/send:
    post:
      consumes:
      - application/json
      description: 发送验证码到手机号/邮箱（identifier=手机号/邮箱），purpose=register/forgot_password
      parameters:
      - description: 发送验证码请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.SendVerifyCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: 发送成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.SendCodeResult'
              type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 发送验证码
      tags:
      - 用户
  /user/info:
    get:
      consumes:
      - application/json
      description: 根据 user_id 查询用户详情，如果不传 user_id 则查询当前登录用户
      parameters:
      - description: 用户ID (不传则查自己)
        format: int64
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.UserDTO'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户信息
      tags:
      - 用户
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录并返回 token（account 支持 username/phone/email；password 或 code 二选一）
      parameters:
      - description: 登录信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/service.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录响应（token + 用户信息）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.LoginResp'
              type: object
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户登录
      tags:
      - 用户
  /user/password:
    post:
      consumes:
      - application/json
      description: 修改当前用户密码（验证码/鉴权由调用层保证）
      parameters:
      - description: 密码修改
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.UpdateUserPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修改用户密码
      tags:
      - 用户
  /user/password/forgot:
    post:
      consumes:
      - application/json
      description: 通过验证码重置密码（identifier 支持 phone/email/username；推荐 phone/email）
      parameters:
      - description: 忘记密码请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/service.ForgotPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 忘记密码
      tags:
      - 用户
  /user/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账号：username + (phone/email 二选一) + password + code
      parameters:
      - description: 注册信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/service.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册
      tags:
      - 用户
  /user/search:
    get:
      consumes:
      - application/json
      description: 按关键字搜索用户（username/nickname/uid），自动排除当前用户
      parameters:
      - description: 搜索关键字
        in: query
        name: keyword
        type: string
      - description: 返回条数
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.UserDTO'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 搜索用户
      tags:
      - 用户
  /user/update:
    post:
      consumes:
      - application/json
      description: 更新当前用户资料（昵称/签名/性别等）
      parameters:
      - description: 更新信息（可选字段）
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/service.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的用户信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.UserDTO'
              type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - 用户
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
  QueryToken:
    description: 用于 WebSocket 等无法传 header 的场景
    in: query
    name: token
    type: apiKey
swagger: "2.0"